RDPP-891: ADI analytics logging should only contain api routes

+++ test/logger.js
@@ -10,15 +10,21 @@ var should = require('should'),
 	request = require('request'),
 	tmpdir = path.join(require('os').tmpdir(), 'test-logger-' + Date.now());
 
-function readFile (filePath, cb) {
+function readFile (filePath, deleteFile, cb) {
+	if (typeof deleteFile === 'function') {
+		cb = deleteFile;
+		deleteFile = true;
+	}
 	fs.readFile(filePath, 'utf8', function (err,data) {
 		if (err) {
 			cb(err);
 			return;
 		}
-		try {
-			fs.unlinkSync(filePath);
-		} catch (e) {}
+		if (deleteFile) {
+			try {
+				fs.unlinkSync(filePath);
+			} catch (e) {}
+		}
 		cb(null, data);
 	});
 }
@@ -181,7 +187,7 @@ describe('ADI logging', function () {
 		fs.emptyDir(tmpdir, done);
 	});
 
-	it('Should log adi logs if singleRequest(transactionLogEnabled) is false', function (callback) {
+	it('RDPP-910: Should log adi logs if singleRequest(transactionLogEnabled) is false', function (callback) {
 		var app = express();
 		_util.findRandomPort(function (err, port) {
 			should(err).be.not.ok;
@@ -192,7 +198,8 @@ describe('ADI logging', function () {
 					logs: tmpdir,
 					logSingleRequest: false,
 					adiLogging: true,
-					name: 'arrowTest'
+					name: 'arrowTest',
+					adiWhitelist: ['/echo']
 				};
 				var logger = index.createExpressLogger(app, loggerConfig);
 				app.use(function (req, resp, next) {
@@ -218,7 +225,7 @@ describe('ADI logging', function () {
 		});
 	});
 
-	it('Should not log arrowPing', function (callback) {
+	it('RDPP-912: Should not log the arrowPing.json healthcheck endpoint', function (callback) {
 		var app = express();
 		_util.findRandomPort(function (err, port) {
 			should(err).be.not.ok;
@@ -229,7 +236,8 @@ describe('ADI logging', function () {
 					logs: tmpdir,
 					logSingleRequest: false,
 					adiLogging: true,
-					name: 'arrowTest'
+					name: 'arrowTest',
+					adiWhitelist: ['/echo']
 				};
 				var logger = index.createExpressLogger(app, loggerConfig);
 				app.use(function (req, resp, next) {
@@ -256,8 +264,7 @@ describe('ADI logging', function () {
 		});
 	});
 
-	// FIXME: See RDPP-891
-	it.skip('Should only log api calls to adi-analytics.log', function (callback) {
+	it('RDPP-911: Should only log items in the whitelist to adi-analytics.log', function (callback) {
 		var app = express();
 		_util.findRandomPort(function (err, port) {
 			should(err).be.not.ok;
@@ -268,7 +275,8 @@ describe('ADI logging', function () {
 					logs: tmpdir,
 					logSingleRequest: false,
 					adiLogging: true,
-					name: 'arrowTest'
+					name: 'arrowTest',
+					adiWhitelist: ['/api']
 				};
 				var logger = index.createExpressLogger(app, loggerConfig);
 				app.use(function (req, resp, next) {
@@ -283,11 +291,11 @@ describe('ADI logging', function () {
 					resp.send({ping: 'pong'});
 					next();
 				});
-				request.get('http://127.0.0.1:' + port + '/echo', function (err, res, body) {
+				request.get('http://127.0.0.1:' + port + '/echo/foo', function (err, res, body) {
 					should(err).not.be.ok;
 					var logPath = path.join(tmpdir, 'adi-analytics.log');
 					should(fs.existsSync(logPath)).be.true;
-					readFile(logPath, function (err, data) {
+					readFile(logPath, false, function (err, data) {
 						should(err).equal(null);
 						should(function () {
 							var logContent = JSON.parse(data);
@@ -320,7 +328,8 @@ describe('ADI logging', function () {
 					logs: tmpdir,
 					logSingleRequest: true,
 					adiLogging: true,
-					name: 'arrowTest'
+					name: 'arrowTest',
+					adiWhitelist: ['/echo']
 				};
 				var logger = index.createExpressLogger(app, loggerConfig);
 				app.use(function (req, resp, next) {
@@ -339,8 +348,8 @@ describe('ADI logging', function () {
 						should(err).equal(null);
 						var logContent = JSON.parse(data);
 						should(logContent.length).not.equal(0);
+						callback();
 					});
-					callback();
 				});
 			});
 		});
@@ -357,7 +366,8 @@ describe('ADI logging', function () {
 					logs: tmpdir,
 					logSingleRequest: true,
 					adiLogging: true,
-					name: 'arrowTest'
+					name: 'arrowTest',
+					adiWhitelist: ['/echo']
 				};
 				var logger = index.createExpressLogger(app, loggerConfig);
 				app.use(function (req, resp, next) {
@@ -377,8 +387,8 @@ describe('ADI logging', function () {
 						var logContent = JSON.parse(data);
 						should(logContent.length).not.equal(0);
 						should(logContent.protocolSrc).equal(port.toString());
+						callback();
 					});
-					callback();
 				});
 
 			});
@@ -396,7 +406,8 @@ describe('ADI logging', function () {
 					logs: tmpdir,
 					logSingleRequest: true,
 					adiLogging: true,
-					name: 'arrowTest'
+					name: 'arrowTest',
+					adiWhitelist: ['/echo']
 				};
 				var logger = index.createExpressLogger(app, loggerConfig);
 				app.use(function (req, resp, next) {
@@ -417,8 +428,8 @@ describe('ADI logging', function () {
 						var logContent = JSON.parse(data);
 						should(logContent.length).not.equal(0);
 						should(logContent.correlationId).equal(res.headers['request-id']);
+						callback();
 					});
-					callback();
 				});
 
 			});
@@ -436,7 +447,8 @@ describe('ADI logging', function () {
 					logs: tmpdir,
 					logSingleRequest: true,
 					adiLogging: true,
-					name: 'arrowTest'
+					name: 'arrowTest',
+					adiWhitelist: ['/echo', '/hundred', '/twoHundred', '/threeHundred', '/fourHundred', '/fiveHundred']
 				};
 				var logger = index.createExpressLogger(app, loggerConfig);
 				app.use(function (req, resp, next) {
@@ -500,10 +512,9 @@ describe('ADI logging', function () {
 						should(threeHundred.status).equal('success');
 						should(fourHundred.status).equal('failure');
 						should(fiveHundred.status).equal('failure');
+						callback();
 					});
-					callback();
 				});
-
 			});
 		});
 	});
@@ -519,7 +530,8 @@ describe('ADI logging', function () {
 					logs: tmpdir,
 					logSingleRequest: true,
 					adiLogging: true,
-					name: 'arrowTest'
+					name: 'arrowTest',
+					adiWhitelist: ['/echo']
 				};
 				var logger = index.createExpressLogger(app, loggerConfig);
 				app.use(function (req, resp, next) {
@@ -540,10 +552,9 @@ describe('ADI logging', function () {
 						var logContent = JSON.parse(data);
 						should(logContent.length).not.equal(0);
 						should(logContent.correlationId).equal(null);
+						callback();
 					});
-					callback();
 				});
-
 			});
 		});
 	});
